Reuse subscription object between Publishers
