+++ Sources/ObservedObject/Property Publishing/Pipeline Factories/PipelineFactory.swift
@@ -1,0 +1,9 @@
+import Combine
+
+protocol PipelineFactory {
+    
+    associatedtype Pipeline: Publisher where Pipeline.Failure == Never
+    
+    func makePipeline() -> Pipeline
+    
+}
